refactored code and separated xrpackage into several modules
